+++ rs/execution_environment/src/hypervisor/tests.rs
@@ -951,14 +951,14 @@ fn ic0_canister_version_returns_correct_value() {
     let expected_ctr: u64 = 1;
     assert_eq!(
         result,
-        WasmResult::Reply((expected_ctr as u64).to_le_bytes().to_vec())
+        WasmResult::Reply(expected_ctr.to_le_bytes().to_vec())
     );
 
     let result = test.ingress(canister_id, "update", ctr.clone()).unwrap();
     let expected_ctr: u64 = 1;
     assert_eq!(
         result,
-        WasmResult::Reply((expected_ctr as u64).to_le_bytes().to_vec())
+        WasmResult::Reply(expected_ctr.to_le_bytes().to_vec())
     );
 
     test.upgrade_canister(canister_id, vec![]).unwrap_err();
@@ -968,7 +968,7 @@ fn ic0_canister_version_returns_correct_value() {
     let expected_ctr: u64 = 2;
     assert_eq!(
         result,
-        WasmResult::Reply((expected_ctr as u64).to_le_bytes().to_vec())
+        WasmResult::Reply(expected_ctr.to_le_bytes().to_vec())
     );
 
     test.uninstall_code(canister_id).unwrap();
@@ -979,7 +979,7 @@ fn ic0_canister_version_returns_correct_value() {
     let expected_ctr: u64 = 4;
     assert_eq!(
         result,
-        WasmResult::Reply((expected_ctr as u64).to_le_bytes().to_vec())
+        WasmResult::Reply(expected_ctr.to_le_bytes().to_vec())
     );
 
     test.uninstall_code(canister_id).unwrap();
@@ -995,7 +995,7 @@ fn ic0_canister_version_returns_correct_value() {
     let expected_ctr: u64 = 6;
     assert_eq!(
         result,
-        WasmResult::Reply((expected_ctr as u64).to_le_bytes().to_vec())
+        WasmResult::Reply(expected_ctr.to_le_bytes().to_vec())
     );
 
     test.reinstall_canister(canister_id, vec![]).unwrap_err();
@@ -1005,7 +1005,7 @@ fn ic0_canister_version_returns_correct_value() {
     let expected_ctr: u64 = 7;
     assert_eq!(
         result,
-        WasmResult::Reply((expected_ctr as u64).to_le_bytes().to_vec())
+        WasmResult::Reply(expected_ctr.to_le_bytes().to_vec())
     );
 
     test.update_freezing_threshold(canister_id, NumSeconds::from(1))
@@ -1014,7 +1014,7 @@ fn ic0_canister_version_returns_correct_value() {
     let expected_ctr: u64 = 8;
     assert_eq!(
         result,
-        WasmResult::Reply((expected_ctr as u64).to_le_bytes().to_vec())
+        WasmResult::Reply(expected_ctr.to_le_bytes().to_vec())
     );
 
     test.canister_update_allocations_settings(canister_id, None, None)
@@ -1023,7 +1023,7 @@ fn ic0_canister_version_returns_correct_value() {
     let expected_ctr: u64 = 9;
     assert_eq!(
         result,
-        WasmResult::Reply((expected_ctr as u64).to_le_bytes().to_vec())
+        WasmResult::Reply(expected_ctr.to_le_bytes().to_vec())
     );
 
     test.canister_update_allocations_settings(canister_id, Some(1000), None)
@@ -1032,7 +1032,7 @@ fn ic0_canister_version_returns_correct_value() {
     let expected_ctr: u64 = 9;
     assert_eq!(
         result,
-        WasmResult::Reply((expected_ctr as u64).to_le_bytes().to_vec())
+        WasmResult::Reply(expected_ctr.to_le_bytes().to_vec())
     );
 
     test.set_controller(canister_id, canister_id.into())
@@ -1041,7 +1041,7 @@ fn ic0_canister_version_returns_correct_value() {
     let expected_ctr: u64 = 10;
     assert_eq!(
         result,
-        WasmResult::Reply((expected_ctr as u64).to_le_bytes().to_vec())
+        WasmResult::Reply(expected_ctr.to_le_bytes().to_vec())
     );
 
     test.uninstall_code(canister_id).unwrap_err();
@@ -1049,7 +1049,7 @@ fn ic0_canister_version_returns_correct_value() {
     let expected_ctr: u64 = 10;
     assert_eq!(
         result,
-        WasmResult::Reply((expected_ctr as u64).to_le_bytes().to_vec())
+        WasmResult::Reply(expected_ctr.to_le_bytes().to_vec())
     );
 }
